Clean code, Replace class Person to Player Engine, Replece class Player to Person, Delete Obj Hand and move to Utility.
